requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=2.1.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
box2d=2.3.2=pypi_0
box2d-py=2.3.5=pypi_0
brotli=1.1.0=hb547adb_1
brotli-bin=1.1.0=hb547adb_1
bzip2=1.0.8=h93a5062_5
c-ares=1.28.1=h93a5062_0
ca-certificates=2024.7.4=hf0a4a13_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.3=pypi_0
certifi=2024.7.4=pyhd8ed1ab_0
charset-normalizer=3.3.2=pypi_0
cloudpickle=3.0.0=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=py310h21239e6_0
cycler=0.12.1=pyhd8ed1ab_0
debugpy=1.8.2=py310hcf9f62a_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
flatbuffers=24.3.25=pypi_0
fonttools=4.53.1=py310ha6dd24b_0
freetype=2.12.1=hadb7bae_2
gast=0.4.0=pypi_0
google-auth=2.32.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.46.3=py310ha26ec5d_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pyhd8ed1ab_0
h5py=3.6.0=nompi_py310hb8bbf05_100
hdf5=1.12.1=nompi_hd9dbc9e_104
idna=3.7=pypi_0
importlib-metadata=8.0.0=pyha770c72_0
importlib_metadata=8.0.0=hd8ed1ab_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.26.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
joblib=1.4.2=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py310hbe9552e_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=py310h38f39d4_1
krb5=1.21.3=h237132a_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=22_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=22_osxarm64_openblas
libclang=18.1.1=pypi_0
libcurl=8.8.0=h7b6f9a7_1
libcxx=17.0.6=h0812c0d_3
libdeflate=1.20=h93a5062_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h93a5062_2
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=22_osxarm64_openblas
libnghttp2=1.58.0=ha4dd798_1
libopenblas=0.3.27=openmp_h517c56d_1
libpng=1.6.43=h091b4b1_0
libprotobuf=3.19.6=hb5ab8b9_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.46.0=hfb93653_0
libssh2=1.11.0=h7a5bd25_0
libtiff=4.6.0=h07db509_3
libwebp-base=1.4.0=h93a5062_0
libxcb=1.15=hf346824_0
libzlib=1.2.13=hfb2fe0b_6
llvm-openmp=18.1.8=hde57baf_0
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=py310hb6292c7_2
matplotlib-base=3.8.4=py310hedb7998_2
matplotlib-inline=0.1.7=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.5=hb89a1cb_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.3.1=hfb2fe0b_2
opt-einsum=3.3.0=pypi_0
packaging=24.1=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py310h81a8c2e_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.2=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
protobuf=3.19.6=py310h0f1eb42_0
psutil=6.0.0=py310ha6dd24b_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pygame=2.1.0=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.2=pyhd8ed1ab_0
python=3.10.14=h2469fbe_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.10=4_cp310
pyvirtualdisplay=3.0=pypi_0
pyzmq=26.0.3=py310h16e08c9_0
readline=8.2=h92ec313_1
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.14.0=pypi_0
setuptools=70.1.1=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
stack_data=0.6.2=pyhd8ed1ab_0
swig=4.2.1=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-deps=2.10.0=0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-macos=2.10.0=pypi_0
tensorflow-metal=0.6.0=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=h5083fa2_1
tornado=6.4.1=py310ha6dd24b_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h0c530f3_0
unicodedata2=15.1.0=py310h2aa6e3c_0
urllib3=2.2.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
wrapt=1.16.0=pypi_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
zeromq=4.3.5=hcc0f68c_4
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=hfb2fe0b_6
zstd=1.5.6=hb46c0d2_0
